fix(mongo): attempt reconnect only on relevant errors
Mongo driver's `handleStoreError` would try reconnection on basically
any `ServerError`, which includes a huge class of failures, e.g. failure
to writing due to exceeding the document size. Reconnecting would not
help in majority of cases there.

This change limits reconnection attempts to network errors and specific
error codes. Reconnection is expensive and shouldn't be done when it
makes no sense.

It also refactors the relevant code by moving reconnect logic to a
separate function.

In addition to this, reverted nilling of the client when closing the
connection to keep the previous behaviour since many functions appear to
expect that `d.client` is always not nil.