fix(mongo): makes disconnect after connect async
MongoDB client disconnection is a blocking operation so it's moved to a
goroutine when re-connecting. This makes the `Connect()` function return
immediately after a successful connect while cleaning up in the
background.

At the same time, ping check is moved to happen before the old client
disconnect to make sure the old client is still used in case the new one
didn't actually connect successfully.

Also, removed the timeout contexts on disconnect so that the timeouts
are consistent via the CSOT (either the default 10s value set on the
client or the connection string option if present).