fix(mongo): set explicit timeout on disconnect calls
Turns out mongo-driver documentation is wrong about CSOT being applied
to all the operations, namely `endSessions` command doesn't set it
implicitly and the connection pool disposal benefits from an explicit
timeout to do it gracefully (otherwise the connections are killed
immediately).

This adds a constant (currently set to 20s) for the disconnect timeout,
which is applied in all the relevant calls.